Rendering continuation
DebugDrawer added to a World.
Inlined suitable functions
Started working on GPUProfiling.
DebugDrawer axes rescaled such that when size is set to 1.0f it will be the same size as a sphere with a radius of 1.0f.
EditorSandbox preinitialization now waits for the graphics creation before creating the sandbox world.
Small bug for material inspector UI that resulted in the reflection manager being allocated from a temporary allocator and not recovered.
Scene now corrects the pacing when using right click (rotation) such that it provides smooth feedback

+++ ECSEngine/src/ECSEngine/Tools/UI/UIDrawer.h
@@ -3694,7 +3694,7 @@ namespace ECSEngine {
 
 			// ------------------------------------------------------------------------------------------------------------------------------------
 			
-			float2 GetAlignedToRight(float x_scale, float target_position = DBL_MAX) const;
+			float2 GetAlignedToRight(float x_scale, float target_position = FLT_MAX) const;
 
 			// ------------------------------------------------------------------------------------------------------------------------------------
 
@@ -3702,7 +3702,7 @@ namespace ECSEngine {
 
 			// ------------------------------------------------------------------------------------------------------------------------------------
 
-			float2 GetAlignedToBottom(float y_scale, float target_position = DBL_MAX) const;
+			float2 GetAlignedToBottom(float y_scale, float target_position = FLT_MAX) const;
 
 			// ------------------------------------------------------------------------------------------------------------------------------------
